refactor(getter): split getter.rs into per-language modules
src/getter.rs had grown to 2673 SLOC — one Getter trait impl per
language (25 of them) plus shared plumbing, 3.3x the 800-line
self-scan limit. Split it per-language exactly like src/metrics/abc
(#969): getter.rs keeps the Getter trait, the shared node_text helper,
the get_operator! / impl_js_family_get_op_type! macros, and the
node_text inline tests; each `impl Getter for <Lang>Code` (plus that
language's private helpers) moves to src/getter/<lang>.rs.

Pure code movement: bodies verbatim, no logic or metric-output change.
The local macros resolve into the child files via textual macro
scoping (the `mod <lang>;` declarations sit after the macro_rules!),
so no macro-export change was needed. Each submodule carries the same
`#![allow(clippy::wildcard_imports, clippy::enum_glob_use)]` +
`use super::*` header as the abc/<lang>.rs precedent. lib.rs is
unchanged (getter is `mod getter;`, not re-exported) — no public API
or STABILITY surface impact.

getter.rs 2673 -> 263 gate SLOC; largest new file is kotlin.rs at 256.
