+++ src/getter/ccomment.rs
@@ -1,0 +1,6 @@
+//! `Getter` implementation for C comments.
+#![allow(clippy::wildcard_imports, clippy::enum_glob_use)]
+
+use super::*;
+
+impl Getter for CcommentCode {}
