+++ src/getter/preproc.rs
@@ -1,0 +1,6 @@
+//! `Getter` implementation for the C preprocessor.
+#![allow(clippy::wildcard_imports, clippy::enum_glob_use)]
+
+use super::*;
+
+impl Getter for PreprocCode {}
